refactor product service responsibilities

+++ services/productService.js
@@ -1,4 +1,5 @@
 const productRepository = require('../repositories/productRepository');
+const AppError = require('../utils/appError');
 
 const productService = {
     async getAllProducts(params) {
@@ -6,7 +7,13 @@ const productService = {
     },
 
     async getProductById(productId) {
-        return productRepository.findById(productId);
+        const product = await productRepository.findById(productId);
+
+        if (!product) {
+        throw new AppError('Produk tidak ditemukan', 404);
+        }
+
+        return product
     },
 
     async createProduct(params) {
@@ -14,11 +21,40 @@ const productService = {
     },
 
     async updateProduct(params) {
+        const {
+            productId,
+            userId,
+            name,
+            price,
+            image,
+        } = params;
+
+        const product = await this.getProductById(productId);
+
+        if (Number(product.user_id) !== Number(userId)) {
+            throw new AppError('Kamu tidak berhak mengedit produk ini',403);
+        }
+
         return productRepository.update(params);
     },
 
-    async deleteProduct(productId) {
-        return productRepository.delete(productId);
+    async deleteProduct({ productId, userId, role }) {
+
+        const product = await this.getProductById(productId);
+
+        if (role === 'admin') {
+            return productRepository.delete(productId)
+        }
+
+        if (role === 'seller') {
+            if (product.user_id !== userId) {
+                throw new AppError('kamu tidak boleh hapus produk ini', 403)
+            }
+
+            return productRepository.delete(productId);
+        }
+
+        throw new AppError('Forbidden', 403);
     },
 
     async getMyProducts(userId) {
